chore(SignUpScreen): screen, button, textfield, header 파일 분리
