+++ app/src/androidTest/java/nextstep/signup/SignUpScreenKtTest.kt
@@ -2,6 +2,7 @@ package nextstep.signup
 
 import androidx.compose.ui.test.junit4.createComposeRule
 import androidx.compose.ui.test.onNodeWithText
+import nextstep.signup.presentation.signup.SignUpHeader
 import org.junit.Rule
 import org.junit.Test
