+++ app/src/main/java/nextstep/signup/presentation/MainActivity.kt
@@ -1,4 +1,4 @@
-package nextstep.signup
+package nextstep.signup.presentation
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -7,6 +7,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Surface
 import androidx.compose.ui.Modifier
+import nextstep.signup.presentation.signup.SignUpScreen
 import nextstep.signup.ui.theme.SignupTheme
 
 class MainActivity : ComponentActivity() {
