+++ app/src/main/java/nextstep/signup/presentation/signup/SignUpButton.kt
@@ -1,0 +1,40 @@
+package nextstep.signup.presentation.signup
+
+import androidx.compose.material3.Button
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.sp
+import nextstep.signup.ui.theme.SignupTheme
+
+@Composable
+fun SignUpButton(
+    modifier: Modifier = Modifier,
+    enable: () -> Boolean = { true }
+) {
+    val enabled = remember { mutableStateOf(true) }
+    Button(
+        onClick = {
+            enabled.value = !enabled.value
+        },
+        enabled = enable(),
+    ) {
+        Text(
+            text = "Sign Up",
+            fontSize = 14.sp,
+            color = Color.White
+        )
+    }
+}
+
+@Preview
+@Composable
+private fun SignUpButtonPreview() {
+    SignupTheme {
+        SignUpButton()
+    }
+}
